app/requirements.txt
flask
google-cloud-firestore
pandas
gunicorn